pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "monarch-pipeline"
version = "0.1.0"
description = "A local data pipeline for Monarch Money — sync accounts, transactions, and budgets to SQLite for analysis and forecasting."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "monarchmoneycommunity",
    "click>=8.0",
    "keyring>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.0",
]

[project.scripts]
monarch-pipeline = "monarch_pipeline.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["monarch_pipeline"]
